+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## Unreleased
+## 1.9.0 (2016-11-17)
+
+We do best effort to support docker versions 1.6.0 - 1.12.1 with this release.
 
 ### Features
 - Build deb, rpm and amazon linux package.